+++ ifood/cadastro/src/main/java/github/com/marcoant007/ifood/repository/PedidosRepository.java
@@ -6,6 +6,6 @@ import github.com.marcoant007.ifood.models.Pedido;
 import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
 
 @ApplicationScoped
-public class PedidosRepository implements PanacheRepositoryBase<Pedido, Integer> {
+public class PedidosRepository implements PanacheRepositoryBase<Pedido, Long> {
     
 }
