+++ ifood/cadastro/src/main/java/github/com/marcoant007/ifood/repository/RestauranteRepository.java
@@ -6,6 +6,6 @@ import github.com.marcoant007.ifood.models.Restaurante;
 import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
 
 @ApplicationScoped
-public class RestauranteRepository implements PanacheRepositoryBase<Restaurante, Integer > {
+public class RestauranteRepository implements PanacheRepositoryBase<Restaurante, Long> {
     
 }
